Creating file oauth.azuread
